Skip TCP test on windows

+++ spec/lib/edr_activity/framework/network/tcp_spec.rb
@@ -1,24 +1,20 @@
 require "socket"
 
 RSpec.describe EDRActivity::Framework::Network do
-  subject(:network_activity) do
-    described_class.new(protocol: :tcp, host: "127.0.0.1", port: @port, message: message)
-  end
+  if Gem.win_platform?
+    xit "sends the message to the destination host and returns the response"
+    xit "returns the size, in bytes, of the message sent"
+    xit "is expected to log activity"
+  else
+    subject(:network_activity) do
+      described_class.new(protocol: :tcp, host: "127.0.0.1", port: @port, message: message)
+    end
 
-  let(:message) { "hello" }
+    let(:message) { "hello" }
 
-  around do |ex|
-    @server = TCPServer.new("127.0.0.1", 0)
-    @port = @server.addr[1]
-    if Gem.win_platform?
-      # Use Process#spawn on windows
-      @server.close # now that we know an open port, close it and reuse it below
-      @pid = Process.spawn(
-        "ruby -e \"require 'socket'; s = TCPServer.new('127.0.0.1', #{@port}); " \
-        "loop { c = s.accept; c.print 'hello there'; c.close; break; }\""
-      )
-    else
-      # Use Process#fork on linux and mac
+    around do |ex|
+      @server = TCPServer.new("127.0.0.1", 0)
+      @port = @server.addr[1]
       @pid = Process.fork do
         loop do
           client = @server.accept
@@ -27,30 +23,23 @@ RSpec.describe EDRActivity::Framework::Network do
           break
         end
       end
+      ex.run
+    ensure
+      @server&.close
+      Process.kill(:KILL, @pid) if @pid
     end
-    ex.run
-  ensure
-    @server&.close
-    @server&.close
-    if @pid
-      begin
-        Process.kill(:KILL, @pid)
-      rescue Errno::ESRCH
-        false
-      end
+
+    it "sends the message to the destination host and returns the response" do
+      expect(network_activity.call.response).to eq("hello there")
     end
-  end
 
-  it "sends the message to the destination host and returns the response" do
-    expect(network_activity.call.response).to eq("hello there")
-  end
+    it "returns the size, in bytes, of the message sent" do
+      expect(network_activity.call.size).to eq(message.bytesize)
+    end
 
-  it "returns the size, in bytes, of the message sent" do
-    expect(network_activity.call.size).to eq(message.bytesize)
-  end
-
-  it "is expected to log activity" do
-    expect(EDRActivity::Framework.logger).to receive(:info)
-    network_activity.call
+    it "is expected to log activity" do
+      expect(EDRActivity::Framework.logger).to receive(:info)
+      network_activity.call
+    end
   end
 end
